fix(api): changed station capacity type from Long to int

See merge request 2024ss-se-pr-group/24ss-se-pr-qse-10!38

+++ backend/src/main/java/at/ac/tuwien/sepr/groupphase/backend/entity/Station.java
@@ -11,7 +11,7 @@ public class Station {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private Long capacity;
+    private int capacity;
 
     public void setId(long id) {
         this.id = id;
@@ -21,7 +21,7 @@ public class Station {
         this.name = name;
     }
 
-    public void setCapacity(Long capacity) {
+    public void setCapacity(int capacity) {
         this.capacity = capacity;
     }
 
@@ -33,7 +33,7 @@ public class Station {
         return name;
     }
 
-    public Long getCapacity() {
+    public int getCapacity() {
         return capacity;
     }
