+++ backend/src/test/java/at/ac/tuwien/sepr/groupphase/backend/basetest/StationTestData.java
@@ -3,7 +3,7 @@ package at.ac.tuwien.sepr.groupphase.backend.basetest;
 public class StationTestData {
 
     public static final String STATION_NAME = "StationName";
-    public static final Long STATION_CAPACITY = 10L;
+    public static final int STATION_CAPACITY = 10;
 
     public static final String BASE_URI = "/api/v1";
     public static final String MESSAGE_BASE_URI = BASE_URI + "/station";
